model an explicitly resumable parsing stack
In preparation for parsing mutually recursive definitions, add support for
partially parsing expressions that contain some (as yet) unknown identifiers.

This means that a parsing operation is no longer a single function call that
returns a parse result, but now consists of an explicit stack that can be
"resumed" in the future.